requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
asyncpg==0.29.0
pytest==7.4.3
httpx==0.25.2

# Advanced features dependencies
redis==5.0.1
pandas==2.1.4
openpyxl==3.1.2
psutil==5.9.6
pyarrow==14.0.1
